Add updated tester.sh and tests with syntax errors
I removed the header from all the base files and fixed a few incorrect
cpsl errors. Some of the tests (like nested_function and
nested_procedure) did not 'forward' declare functions. Also one of the
tests in the for loop was not pre-declaring the variable i. I think you
have to declare the variable so I changed that.